// src/components/types/color.ts
export enum AnyColor {
    Gold1 = "gold1",
    Gold2 = "gold2",
    Gold3 = "gold3",
    Gold4 = "gold4",
    Gold5 = "gold5",
    Blue1 = "blue1",
    Blue2 = "blue2",
    Blue3 = "blue3",
    Blue4 = "blue4",
    Blue5 = "blue5",
    Red1 = "red1",
    Red2 = "red2",
    Red3 = "red3",
    Red4 = "red4",
    Red5 = "red5",
    Green1 = "green1",
    Green2 = "green2",
    Green3 = "green3",
    Green4 = "green4",
    Green5 = "green5",
    Gray1 = "gray1",
    Gray2 = "gray2",
    Gray3 = "gray3",
    Gray4 = "gray4",
    Gray5 = "gray5",
    Gray6 = "gray6",
    Gray7 = "gray7",
    Gray8 = "gray8",
    Gray9 = "gray9",
    Gray10 = "gray10",
    White = "white",

    Brand = "gold3",
    Primary = "blue3",
    Secondary = "gray8",
    Error = "red3",
    Success = "green3"
}